+++ packages/e2e-vite/src/router.ts
@@ -14,6 +14,7 @@ import { TestUseDefaults } from './fixtures/use-defaults.vine'
 import { VaporTestContainer } from './fixtures/vapor-interop.vine'
 import { TestVibe } from './fixtures/vibe.vine'
 import { TestVineEmitsPage } from './fixtures/vine-emits.vine'
+import { TestVineModelModifiers } from './fixtures/vine-model-modifiers.vine'
 import { TestVineModel } from './fixtures/vine-model.vine'
 import { TestVinePropPage } from './fixtures/vine-prop.vine'
 import { TestVineSlots } from './fixtures/vine-slots.vine'
@@ -32,6 +33,7 @@ const routes = [
   { path: '/vibe', component: TestVibe },
   { path: '/use-defaults', component: TestUseDefaults },
   { path: '/vine-model', component: TestVineModel },
+  { path: '/vine-model-modifiers', component: TestVineModelModifiers },
   { path: '/vine-emits', component: TestVineEmitsPage },
   { path: '/vine-validators', component: TestVineValidatorsPage },
   { path: '/todo-list', component: TodoList },