+++ packages/tsc/CHANGELOG.md
@@ -1,5 +1,13 @@
 # vue-vine-tsc
 
+## 1.7.24
+
+### Patch Changes
+
+- Support vineModel destructure as array pattern and get modifiers.
+- Updated dependencies
+  - @vue-vine/language-service@1.7.24
+
 ## 1.7.23
 
 ### Patch Changes